SearchDataTransformer: centralize allResultsForVertical parsing (#786)
* SearchDataTransformer: centralize allResultsForVertical parsing

This commit moves _formResponseFromAllResultsForVertical from models/verticalresults.js
up one level into searchdatatransformer, so that other models for the same set of request,
like DYNAMIC_FILTERS, do not have to redo this parsing. It also adds resultsContext into data.response as well as the Section model, so we can stop treating it like a special flower (even though it is in the sense that it is currently the only property to not come directly from the response, however we don't want to keep tacking on additional parameters)

TEST=manual, auto
checked that can run searches and apply facets for both normal and no results.
